+++ Back_End/server.js
@@ -12,7 +12,6 @@ const cors = require('cors');
 
 //라우팅
 const home = require('./routes');
-const db = require('./routes/db');
 
 //웹세팅
 app.use(bodyParser.json());
@@ -20,7 +19,6 @@ app.use(bodyParser.urlencoded({ extended: true }));
 app.use(cors());
 
 app.use("/", home);
-app.use("/db", db);
 
 //연결
 app.listen(config.get('server.port'), () => { // 서버 연결
